Broke my promise, but trying to make it right.
- Duh! `paths` & `logger` are obviously WIP and need to live in `exp/`.
- Decoupled `logger` from `paths.
